Extract turbine extension
